hailscout-api/pyproject.toml
[tool.poetry]
name = "hailscout-api"
version = "0.1.0"
description = "FastAPI service for HailScout hail tracking platform"
authors = ["HailScout Team <dev@hailscout.com>"]
license = "Proprietary"
readme = "README.md"
packages = [{include = "hailscout_api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = {version = "^0.24.0", extras = ["standard"]}
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
geoalchemy2 = "^0.14.1"
shapely = "^2.0.2"
requests = "^2.31.0"
httpx = "^0.25.2"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
cryptography = "^41.0.7"
structlog = "^24.1.0"
python-json-logger = "^2.0.7"
PyJWT = "^2.8.1"
sentry-sdk = "^1.40.0"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
mypy = "^1.7.1"
ruff = "^0.1.11"
black = "^23.12.0"
pytest-env = "^1.1.1"

[tool.poetry.group.test.dependencies]
httpx = "^0.25.2"
pytest-mock = "^3.12.0"

[tool.mypy]
python_version = "3.12"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "alembic.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "geoalchemy2.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "cfgrib.*"
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific
]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "ARG001", # Unused function argument (common in callbacks)
]

[tool.ruff.isort]
force-single-line = false
known-first-party = ["hailscout_api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "--strict-markers -v"
markers = [
    "integration: marks tests as integration (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
